게시판 타입을 @PathVariable로 받게 변경하여 불필요한 중복코드 제거

+++ server/src/main/java/community/board/controller/BoardController.java
@@ -5,6 +5,7 @@ import community.board.dto.UploadDto;
 import community.board.entity.Board;
 import community.board.entity.UploadFile;
 import community.board.service.S3Service;
+import community.exception.ExceptionCode;
 import community.member.entity.Member;
 import community.member.service.MemberService;
 import community.type.SearchType;
@@ -40,40 +41,24 @@ public class BoardController {
     private final S3Service s3Service;
     private final MemberService memberService;
 
-    @PostMapping("/free")
-    public ResponseEntity<?> createFreeBoard(@ModelAttribute MultipartFile[] files,
-                                             @Valid @ModelAttribute BoardDto.Post boardPostDto) throws Exception {
+    @PostMapping("/{board-type}")
+    public ResponseEntity<?> createBoard(@ModelAttribute MultipartFile[] files,
+                                         @Valid @ModelAttribute BoardDto.Post boardPostDto,
+                                         @PathVariable("board-type") String boardType) throws Exception {
         Board board = boardMapper.boardPostToBoard(boardPostDto);
         board.setMember(memberService.findVerifiedMember(boardPostDto.getMemberId()));
-        Board boardCreate = boardService.createBoard(board, Board.KindOfBoard.FREE_BOARD); //게시판 종류를 입력
+        Board boardCreate;
 
-        List<UploadFile> uploadFiles = s3Service.uploadFiles(files, boardCreate); // aws s3업로드
-        List<UploadDto> uploadResponse = boardMapper.uploadFilesToUploadDtoList(uploadFiles); //업로드 dto리스트를 생성
-
-        BoardDto.TotalPageResponse response = boardMapper.boardToBoardTotalPageResponse(boardCreate, uploadResponse);//게시글 dto에 업로드 dto담아주기
-        return new ResponseEntity<>(new SingleResponseDto<>(response), HttpStatus.CREATED);
-    }
-
-    @PostMapping("/eco")
-    public ResponseEntity<?> createEcoBoard(@ModelAttribute MultipartFile[] files,
-                                            @Valid @ModelAttribute BoardDto.Post boardPostDto) throws Exception {
-        Board board = boardMapper.boardPostToBoard(boardPostDto);
-        board.setMember(memberService.findVerifiedMember(boardPostDto.getMemberId()));
-        Board boardCreate = boardService.createBoard(board, Board.KindOfBoard.ECO_REVIEW); //게시판 종류를 입력
-
-        List<UploadFile> uploadFiles = s3Service.uploadFiles(files, boardCreate); // aws s3업로드
-        List<UploadDto> uploadResponse = boardMapper.uploadFilesToUploadDtoList(uploadFiles); //업로드 dto리스트를 생성
-
-        BoardDto.TotalPageResponse response = boardMapper.boardToBoardTotalPageResponse(boardCreate, uploadResponse);//게시글 dto에 업로드 dto담아주기
-        return new ResponseEntity<>(new SingleResponseDto<>(response), HttpStatus.CREATED);
-    }
-
-    @PostMapping("/green")
-    public ResponseEntity<?> createGreenBoard(@ModelAttribute MultipartFile[] files,
-                                              @Valid @ModelAttribute BoardDto.Post boardPostDto) throws Exception {
-        Board board = boardMapper.boardPostToBoard(boardPostDto);
-        board.setMember(memberService.findVerifiedMember(boardPostDto.getMemberId()));
-        Board boardCreate = boardService.createBoard(board, Board.KindOfBoard.GREEN_ACTIVE); //게시판 종류를 입력
+        //Todo : 중복코드 제거를 위해 3개의 요청을 1개의 요청으로 묶을 수 있다.
+        if(boardType.equals("eco")) {
+            boardCreate = boardService.createBoard(board, Board.KindOfBoard.FREE_BOARD); //게시판 종류를 입력
+        }
+        else if(boardType.equals("free")) {
+            boardCreate = boardService.createBoard(board, Board.KindOfBoard.ECO_REVIEW);
+        }
+        else {
+            boardCreate = boardService.createBoard(board, Board.KindOfBoard.GREEN_ACTIVE);
+        }
 
         List<UploadFile> uploadFiles = s3Service.uploadFiles(files, boardCreate); // aws s3업로드
         List<UploadDto> uploadResponse = boardMapper.uploadFilesToUploadDtoList(uploadFiles); //업로드 dto리스트를 생성
@@ -87,6 +72,7 @@ public class BoardController {
                                          @ModelAttribute MultipartFile[] files,
                                          @ModelAttribute BoardDto.Patch boardPatchDto) throws Exception {
         Board board = boardMapper.boardPatchToBoard(boardPatchDto);
+        // DB에서 memeber를 찾음
         board.setBoardId(boardId);
         Board updateBoard = boardService.updateBoard(board);
 
@@ -98,44 +84,25 @@ public class BoardController {
     }
 
     /*검색 및 전체조회*/
-    @GetMapping("/free") //부분검색 //http://localhost:8080/boards?searchType=CONTENTS&searchValue=검색어&page=&size
-    public ResponseEntity<?> searchFreeBoards(@RequestParam(required = false) SearchType searchType,//required = false - 선택적 파라미터
-                                              @RequestParam(required = false) String searchValue,
-                                              @PageableDefault(size = 100, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) throws Exception //페이지 기본값
+    @GetMapping("/{board-type}") //부분검색 //http://localhost:8080/boards?searchType=CONTENTS&searchValue=검색어&page=&size
+    public ResponseEntity<?> searchBoards(@RequestParam(required = false) SearchType searchType,//required = false - 선택적 파라미터
+                                          @RequestParam(required = false) String searchValue,
+                                          @PageableDefault(size = 100, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable,
+                                          @PathVariable("board-type") String boardType) throws Exception //페이지 기본값
     {
-        boardService.userAndLevelUpdate(); //작성자들의 레벨 및 이름 업데이트
 
-        Page<Board> boardPage = boardService.findBoards(Board.KindOfBoard.FREE_BOARD, searchType, searchValue, pageable);
-        List<Board> boards = boardPage.getContent();
+        Page<Board> boardPage;
 
-        List<BoardDto.TotalPageListResponse> response = boardMapper.boardToBoardListResponse(boards);
-        return new ResponseEntity<>(response, HttpStatus.OK);
-    }
+        if(boardType.equals("eco")) {
+            boardPage = boardService.findBoards(Board.KindOfBoard.ECO_REVIEW, searchType, searchValue, pageable);
+        }
+        else if(boardType.equals("free")) {
+            boardPage = boardService.findBoards(Board.KindOfBoard.FREE_BOARD, searchType, searchValue, pageable);
+        }
+        else {
+            boardPage = boardService.rankBoards(Board.KindOfBoard.GREEN_ACTIVE, pageable);
+        }
 
-    /*검색 및 전체조회*/
-    @GetMapping("/eco") //부분검색 //http://localhost:8080/boards?searchType=CONTENTS&searchValue=검색어&page=&size
-    public ResponseEntity<?> searchEcoBoards(@RequestParam(required = false) SearchType searchType,//required = false - 선택적 파라미터
-                                             @RequestParam(required = false) String searchValue,
-                                             @PageableDefault(size = 100, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) throws Exception //페이지 기본값
-    {
-        boardService.userAndLevelUpdate(); //작성자들의 레벨 및 이름 업데이트 /검색하는사람의 토큰값으로 확인가능
-
-        Page<Board> boardPage = boardService.findBoards(Board.KindOfBoard.ECO_REVIEW, searchType, searchValue, pageable);
-        List<Board> boards = boardPage.getContent();
-
-        List<BoardDto.TotalPageListResponse> response = boardMapper.boardToBoardListResponse(boards);
-        return new ResponseEntity<>(response, HttpStatus.OK);
-    }
-
-    /*검색 및 전체조회*/
-    @GetMapping("/green") //부분검색 //http://localhost:8080/boards?searchType=CONTENTS&searchValue=검색어&page=&size
-    public ResponseEntity<?> searchGreenBoards(@RequestParam(required = false) SearchType searchType,//required = false - 선택적 파라미터
-                                               @RequestParam(required = false) String searchValue,
-                                               @PageableDefault(size = 100, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) throws Exception //페이지 기본값
-    {
-        boardService.userAndLevelUpdate(); //작성자들의 레벨 및 이름 업데이트
-
-        Page<Board> boardPage = boardService.findBoards(Board.KindOfBoard.GREEN_ACTIVE, searchType, searchValue, pageable);
         List<Board> boards = boardPage.getContent();
 
         List<BoardDto.TotalPageListResponse> response = boardMapper.boardToBoardListResponse(boards);
@@ -143,30 +110,21 @@ public class BoardController {
     }
 
     /*추천 게시판*/
-    @GetMapping("/rankFreeBoards")
-    public ResponseEntity<?> rankFreeBoards(@PageableDefault(size = 5, sort = "likeCount", direction = Sort.Direction.DESC) Pageable pageable) //페이지 기본값
+    @GetMapping("/rank+{board-type}")
+    public ResponseEntity<?> rankBoards(@PageableDefault(size = 5, sort = "likeCount", direction = Sort.Direction.DESC) Pageable pageable,
+                                        @PathVariable("board-type") String boardType) //페이지 기본값
     {
-        Page<Board> boardPage = boardService.rankBoards(Board.KindOfBoard.FREE_BOARD, pageable);
-        List<Board> boards = boardPage.getContent();
-        List<BoardDto.RankResponse> response = boardMapper.boardToBoardRankListResponse(boards);
-        return new ResponseEntity<>(response, HttpStatus.OK);
-    }
+        Page<Board> boardPage;
 
-    /*추천 게시판*/
-    @GetMapping("/rankEcoBoards")
-    public ResponseEntity<?> rankEcoBoards(@PageableDefault(size = 5, sort = "likeCount", direction = Sort.Direction.DESC) Pageable pageable) //페이지 기본값
-    {
-        Page<Board> boardPage = boardService.rankBoards(Board.KindOfBoard.ECO_REVIEW, pageable);
-        List<Board> boards = boardPage.getContent();
-        List<BoardDto.RankResponse> response = boardMapper.boardToBoardRankListResponse(boards);
-        return new ResponseEntity<>(response, HttpStatus.OK);
-    }
-
-    /*추천 게시판*/
-    @GetMapping("/rankGreenBoards")
-    public ResponseEntity<?> rankGreenBoards(@PageableDefault(size = 5, sort = "likeCount", direction = Sort.Direction.DESC) Pageable pageable) //페이지 기본값
-    {
-        Page<Board> boardPage = boardService.rankBoards(Board.KindOfBoard.GREEN_ACTIVE, pageable);
+        if(boardType.equals("ecoBoard")) {
+            boardPage = boardService.rankBoards(Board.KindOfBoard.ECO_REVIEW, pageable);
+        }
+        else if(boardType.equals("freeBoard")) {
+            boardPage = boardService.rankBoards(Board.KindOfBoard.FREE_BOARD, pageable);
+        }
+        else {
+            boardPage = boardService.rankBoards(Board.KindOfBoard.GREEN_ACTIVE, pageable);
+        }
         List<Board> boards = boardPage.getContent();
         List<BoardDto.RankResponse> response = boardMapper.boardToBoardRankListResponse(boards);
         return new ResponseEntity<>(response, HttpStatus.OK);
@@ -178,7 +136,6 @@ public class BoardController {
         boardService.updateViewCount(boardId); // 조회수 증가
         Board board = boardService.findBoard(boardId);
 
-//        boardLikeService.BoardLikeStatus(memberId, board); //게시글에대한 좋아요상태 업데이트
 
         List<UploadFile> uploadFiles = s3Service.uploadFiles(null, board); // aws s3업로드
         List<UploadDto> uploadResponse = boardMapper.uploadFilesToUploadDtoList(uploadFiles); //업로드 dto생성
@@ -201,8 +158,16 @@ public class BoardController {
 
     @DeleteMapping("/{board-id}")
     public ResponseEntity<?> deleteBoard(@PathVariable("board-id") @Positive long boardId) {
-        boardService.deleteBoard(boardId);
-        return ResponseEntity.noContent().build();
+        Board findBoard = boardService.findBoardById(boardId);
+        long findMemberId = findBoard.getMember().getMemberId();
+
+        Member loginMember = loginMemberFindByToken();
+        if(findMemberId == loginMember.getMemberId()) {
+            boardService.deleteBoard(boardId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return new ResponseEntity<>(new SingleResponseDto<>(ExceptionCode.BOARD_NOT_FOUND),HttpStatus.ACCEPTED);
+        }
     }
 
     @PostMapping("/{board-id}/Like")
